refactor(persistence): split derived-index ports from their adapters
`index/` is the package outer layers depend on, and it held both the
abstraction and 515 lines implementing it -- the LanceDB adapter and the
Milvus lifecycle shim sat next to the protocols they satisfy. Anyone
opening it to add a third backend had to work out which files were the
contract by reading them.

The adapters move to `persistence/backends/`. `index/` keeps the ports,
the portable schema, the router and the facade; nothing else changes,
and no behaviour does -- the facade re-exports the same names, so every
caller outside `persistence` is untouched.

Two contracts make the split hold, because moving files does not stop
anyone moving them back:

* Outer layers may not import `backends.**`, matching the rule already
  covering the other storage subpackages.
* `index.protocols`, `index.schema` and the shared `predicate` module may
  not import any concrete backend. The facade and the router resolve a
  configured backend and legitimately do; the port definitions must not,
  or "portable" stops being a checkable claim. Verified by breaking each
  rule in turn and confirming import-linter reports it.

Six test modules referenced the adapters by their old path and move with
them.
