+++ tests/unit/test_memory/test_search/test_manager.py
@@ -393,7 +393,7 @@ async def test_user_keyword_filters_compile_pinned_owner() -> None:
     )
     await mgr.search(_user_req())
     assert recaller.last_where is not None
-    from everos.infra.persistence.index.lancedb import render_predicate
+    from everos.infra.persistence.backends.lancedb import render_predicate
 
     rendered = render_predicate(recaller.last_where)
     assert "owner_id = 'alice'" in rendered
